fix: move SetFlagValue function to flag_value.ts for better organization
